Add support for drawing images to the frame and apply to de_jong_ifs
This is fairly complex to do, as it requires writing out the image as a
texture. Also, some DPI weirdness still needs to be resolved. For
example, on my Windows laptop the window API returns my DPI as 1.25.
However, to actually scale the image and framebuffer to fill the window
they need to be scaled up by a factor of 2. This needs more attention to
see what the correct behavior to follow is.